+++ packages/core/src/loaders/imageLoader.ts
@@ -1,15 +1,51 @@
 import cache from '../cache/cache';
 import Events from '../enums/Events';
 import eventTarget from '../eventTarget';
-import { triggerEvent } from '../utilities';
-import { IImage, ImageLoaderFn, IImageLoadObject, EventTypes } from '../types';
+import {
+  genericMetadataProvider,
+  getBufferConfiguration,
+  triggerEvent,
+  uuidv4,
+} from '../utilities';
+import {
+  IImage,
+  ImageLoaderFn,
+  IImageLoadObject,
+  EventTypes,
+  Point2,
+  Point3,
+  Mat3,
+  PixelDataTypedArrayString,
+  PixelDataTypedArray,
+} from '../types';
 import imageLoadPoolManager from '../requestPool/imageLoadPoolManager';
+import { metaData } from '../';
 
 export interface ImageLoaderOptions {
   priority: number;
   requestType: string;
   additionalDetails?: Record<string, unknown>;
 }
+
+interface DerivedImageOptions {
+  imageId?: string;
+  targetBufferType?: PixelDataTypedArrayString;
+}
+
+interface DerivedImages {
+  imageIds: Array<string>;
+  promises: Array<Promise<IImage>>;
+}
+
+interface LocalImageOptions {
+  scalarData?: PixelDataTypedArray;
+  targetBufferType?: PixelDataTypedArrayString;
+  dimensions?: Point2;
+  spacing?: Point3;
+  origin?: Point3;
+  direction?: Mat3;
+}
+
 /**
  * This module deals with ImageLoaders, loading images and caching images
  */
@@ -181,6 +217,170 @@ export function loadAndCacheImages(
   });
 
   return allPromises;
+}
+
+/**
+ * Loads an image given an imageId and optional priority and returns a promise
+ * which will resolve to the loaded image object or fail if an error occurred.
+ * The image is stored in the cache.
+ *
+ * @param referencedImageId -  A Cornerstone Image Object's imageId
+ * @param options - Options to be passed to the Image Loader
+ *
+ * @returns Image Loader Object
+ */
+export function createAndCacheDerivedImage(
+  referencedImageId: string,
+  options: DerivedImageOptions = {},
+  preventCache = false
+): Promise<IImage> {
+  if (referencedImageId === undefined) {
+    throw new Error(
+      'createAndCacheDerivedImage: parameter imageId must not be undefined'
+    );
+  }
+
+  if (options.imageId === undefined) {
+    options.imageId = `derived:${uuidv4()}`;
+  }
+
+  const imagePlaneModule = metaData.get('imagePlaneModule', referencedImageId);
+
+  const length = imagePlaneModule.rows * imagePlaneModule.columns;
+
+  const { TypedArrayConstructor } = getBufferConfiguration(
+    options.targetBufferType,
+    length
+  );
+
+  const imageScalarData = new TypedArrayConstructor(length);
+  const derivedImageId = options.imageId;
+
+  ['imagePixelModule', 'imagePlaneModule', 'generalSeriesModule'].forEach(
+    (type) => {
+      genericMetadataProvider.add(derivedImageId, {
+        type,
+        metadata: metaData.get(type, referencedImageId),
+      });
+    }
+  );
+
+  const localImage = createAndCacheLocalImage(
+    { scalarData: imageScalarData },
+    options.imageId,
+    true
+  );
+
+  const imageLoadObject = {
+    promise: Promise.resolve(localImage),
+  };
+
+  if (!preventCache) {
+    cache.putImageLoadObject(derivedImageId, imageLoadObject);
+  }
+  return imageLoadObject.promise;
+}
+
+/**
+ * Load and cache a list of imageIds
+ *
+ * @param referencedImageIds - list of imageIds
+ * @param getDerivedImageId - optional function to generate derived imageId name however you want
+ */
+export function createAndCacheDerivedImages(
+  referencedImageIds: Array<string>,
+  getDerivedImageId?: (referencedImageId: string) => string
+): DerivedImages {
+  if (referencedImageIds?.length === 0) {
+    throw new Error(
+      'createAndCacheDerivedImages: parameter imageIds must be list of image Ids'
+    );
+  }
+
+  const derivedImageIds = [];
+  const allPromises = referencedImageIds.map((referencedImageId, index) => {
+    const options: DerivedImageOptions = {
+      imageId: getDerivedImageId
+        ? getDerivedImageId(referencedImageId)
+        : `derived:${uuidv4()}`,
+    };
+    derivedImageIds.push(options.imageId);
+    return createAndCacheDerivedImage(referencedImageId, options);
+  });
+
+  return { imageIds: derivedImageIds, promises: allPromises };
+}
+
+export function createAndCacheLocalImage(
+  options: LocalImageOptions,
+  imageId: string,
+  preventCache = false
+): IImage {
+  const imagePlaneModule = metaData.get('imagePlaneModule', imageId);
+
+  const length = imagePlaneModule.rows * imagePlaneModule.columns;
+
+  const image = {
+    imageId: imageId,
+    intercept: 0,
+    windowCenter: 0,
+    windowWidth: 0,
+    color: false,
+    numComps: 1,
+    slope: 1,
+    minPixelValue: 0,
+    maxPixelValue: 255,
+    voiLUTFunction: undefined,
+    rows: imagePlaneModule.rows,
+    columns: imagePlaneModule.columns,
+    getCanvas: undefined, // todo: which canvas?
+    height: imagePlaneModule.rows,
+    width: imagePlaneModule.columns,
+    rgba: undefined, // todo: how
+    columnPixelSpacing: imagePlaneModule.columnPixelSpacing,
+    rowPixelSpacing: imagePlaneModule.rowPixelSpacing,
+    invert: false,
+  } as IImage;
+
+  if (options.scalarData) {
+    const imageScalarData = options.scalarData;
+
+    if (
+      !(
+        imageScalarData instanceof Uint8Array ||
+        imageScalarData instanceof Float32Array ||
+        imageScalarData instanceof Uint16Array ||
+        imageScalarData instanceof Int16Array
+      )
+    ) {
+      throw new Error(
+        'To use createLocalVolume you should pass scalarData of type Uint8Array, Uint16Array, Int16Array or Float32Array'
+      );
+    }
+
+    image.sizeInBytes = imageScalarData.byteLength;
+    image.getPixelData = () => imageScalarData;
+  } else {
+    const { numBytes, TypedArrayConstructor } = getBufferConfiguration(
+      options.targetBufferType,
+      length
+    );
+
+    const imageScalarData = new TypedArrayConstructor(length);
+
+    image.sizeInBytes = numBytes;
+    image.getPixelData = () => imageScalarData;
+  }
+
+  const imageLoadObject = {
+    promise: Promise.resolve(image),
+  };
+
+  if (!preventCache) {
+    cache.putImageLoadObject(image.imageId, imageLoadObject);
+  }
+
+  return image;
 }
 
 /**