+++ packages/tools/src/tools/displayTools/Labelmap/validateRepresentationData.ts
@@ -1,6 +1,9 @@
 import { SegmentationPublicInput } from '../../../types/SegmentationStateTypes';
 import { cache } from '@cornerstonejs/core';
-import { LabelmapSegmentationData } from '../../../types/LabelmapTypes';
+import type {
+  LabelmapSegmentationData,
+  LabelmapSegmentationDataVolume,
+} from '../../../types/LabelmapTypes';
 
 function validate(segmentationInput: SegmentationPublicInput): void {
   if (!segmentationInput.representation.data) {
@@ -12,18 +15,21 @@ function validate(segmentationInput: SegmentationPublicInput): void {
   const representationData = segmentationInput.representation
     .data as LabelmapSegmentationData;
 
-  if (!representationData.volumeId) {
-    throw new Error(
-      'The segmentationInput.representationData.volumeId is undefined, please provide a valid representationData.volumeId'
+  if ('volumeId' in representationData) {
+    // volumetric labelmap
+    const cachedVolume = cache.getVolume(
+      (representationData as LabelmapSegmentationDataVolume).volumeId
     );
-  }
 
-  const cachedVolume = cache.getVolume(representationData.volumeId);
-
-  if (!cachedVolume) {
-    throw new Error(
-      `volumeId of ${representationData.volumeId} not found in cache, you should load and cache volume before adding segmentation`
-    );
+    if (!cachedVolume) {
+      throw new Error(
+        `volumeId of ${
+          (representationData as LabelmapSegmentationDataVolume).volumeId
+        } not found in cache, you should load and cache volume before adding segmentation`
+      );
+    }
+  } else {
+    // I don't think we need this check since there is no guarantee that the stack is cached.
   }
 }