+++ packages/tools/src/tools/segmentation/CircleScissorsTool.ts
@@ -1,4 +1,4 @@
-import { cache, getEnabledElement, StackViewport } from '@cornerstonejs/core';
+import { cache, getEnabledElement } from '@cornerstonejs/core';
 import type { Types } from '@cornerstonejs/core';
 
 import { BaseTool } from '../base';
@@ -25,7 +25,12 @@ import {
   config as segmentationConfig,
 } from '../../stateManagement/segmentation';
 import { getSegmentation } from '../../stateManagement/segmentation/segmentationState';
-import { LabelmapSegmentationData } from '../../types/LabelmapTypes';
+import {
+  LabelmapSegmentationData,
+  LabelmapSegmentationDataStack,
+  LabelmapSegmentationDataVolume,
+} from '../../types/LabelmapTypes';
+import { isVolumeSegmentation } from './strategies/utils/stackVolumeCheck';
 
 /**
  * Tool for manipulating segmentation data by drawing a circle. It acts on the
@@ -38,9 +43,12 @@ class CircleScissorsTool extends BaseTool {
   static toolName;
   editData: {
     annotation: any;
-    segmentation: any;
     segmentIndex: number;
-    segmentationId: string;
+    //
+    volumeId: string;
+    referencedVolumeId: string;
+    imageIdReferenceMap: Map<string, string>;
+    //
     segmentsLocked: number[];
     segmentColor: [number, number, number, number];
     viewportIdsToRender: string[];
@@ -49,6 +57,7 @@ class CircleScissorsTool extends BaseTool {
     newAnnotation?: boolean;
     hasMoved?: boolean;
     centerCanvas?: Array<number>;
+    segmentationRepresentationUID?: string;
   } | null;
   isDrawing: boolean;
   isHandleOutsideImage: boolean;
@@ -116,8 +125,13 @@ class CircleScissorsTool extends BaseTool {
     const { representationData } = getSegmentation(segmentationId);
 
     // Todo: are we going to support contour editing with rectangle scissors?
-    const { volumeId } = representationData[type] as LabelmapSegmentationData;
-    const segmentation = cache.getVolume(volumeId);
+    const labelmapData = representationData[type];
+
+    if (!labelmapData) {
+      throw new Error(
+        'No labelmap data found for the active segmentation, create one before using scissors tool'
+      );
+    }
 
     // Todo: Used for drawing the svg only, we might not need it at all
     const annotation = {
@@ -145,7 +159,6 @@ class CircleScissorsTool extends BaseTool {
 
     this.editData = {
       annotation,
-      segmentation,
       centerCanvas: canvasPos,
       segmentIndex,
       segmentationId,
@@ -156,7 +169,27 @@ class CircleScissorsTool extends BaseTool {
       movingTextBox: false,
       newAnnotation: true,
       hasMoved: false,
-    };
+      segmentationRepresentationUID,
+    } as any;
+
+    if (isVolumeSegmentation(labelmapData as LabelmapSegmentationData)) {
+      const { volumeId } = labelmapData as LabelmapSegmentationDataVolume;
+      const segmentation = cache.getVolume(volumeId);
+
+      this.editData = {
+        ...this.editData,
+        volumeId,
+        referencedVolumeId: segmentation.referencedVolumeId,
+      };
+    } else {
+      const { imageIdReferenceMap } =
+        labelmapData as LabelmapSegmentationDataStack;
+
+      this.editData = {
+        ...this.editData,
+        imageIdReferenceMap,
+      };
+    }
 
     this._activateDraw(element);
 
@@ -221,15 +254,7 @@ class CircleScissorsTool extends BaseTool {
     const eventDetail = evt.detail;
     const { element } = eventDetail;
 
-    const {
-      annotation,
-      newAnnotation,
-      hasMoved,
-      segmentation,
-      segmentIndex,
-      segmentsLocked,
-      segmentationId,
-    } = this.editData;
+    const { annotation, newAnnotation, hasMoved } = this.editData;
     const { data } = annotation;
     const { viewPlaneNormal, viewUp } = annotation.metadata;
 
@@ -244,25 +269,17 @@ class CircleScissorsTool extends BaseTool {
     resetElementCursor(element);
 
     const enabledElement = getEnabledElement(element);
-    const { viewport } = enabledElement;
-
-    this.editData = null;
-    this.isDrawing = false;
-
-    if (viewport instanceof StackViewport) {
-      throw new Error('Not implemented yet');
-    }
 
     const operationData = {
+      ...this.editData,
       points: data.handles.points,
-      volume: segmentation,
-      segmentIndex,
-      segmentsLocked,
       viewPlaneNormal,
-      segmentationId,
       viewUp,
       strategySpecificConfiguration: {},
     };
+
+    this.editData = null;
+    this.isDrawing = false;
 
     this.applyActiveStrategy(enabledElement, operationData);
   };