+++ packages/tools/src/tools/segmentation/strategies/eraseSphere.ts
@@ -1,18 +1,10 @@
 import type { Types } from '@cornerstonejs/core';
 
 import { fillInsideSphere } from './fillSphere';
+import { LabelmapToolOperationData } from '../../../types';
 
-type OperationData = {
+type OperationData = LabelmapToolOperationData & {
   points: [Types.Point3, Types.Point3, Types.Point3, Types.Point3];
-  imageVolume: Types.IImageVolume;
-  volume: Types.IImageVolume;
-  segmentIndex: number;
-  segmentationId: string;
-  segmentsLocked: number[];
-  viewPlaneNormal: Types.Point3;
-  viewUp: Types.Point3;
-  constraintFn: () => boolean;
-  strategySpecificConfiguration: any;
 };
 
 export function eraseInsideSphere(